+++ src/main/java/com/ircclouds/irc/api/dcc/interfaces/IDCCSendResult.java
@@ -3,8 +3,8 @@ package com.ircclouds.irc.api.dcc.interfaces;
 public interface IDCCSendResult
 {
 	int getNumberOfAcksReceived();
-	
-	int totalBytesSent();
+
+	long totalBytesSent();
 	
 	long totalTime();
 }
